# blackjack-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blackjack API",
        "description": "This is a simple Blackjack API. For more information, please visit the developer's portfolio at [Ignacio Albiol's Portfolio](https://ignacioalbiol.es/).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Blackjack API Server"
        }
    ],
    "paths": {
...
        "/players/ranking": {
            "get": {
                "tags": [
                    "Ranking"
                ],
                "summary": "Get the ranking of all players",
                "description": "This endpoint returns the ranking of all players.",
                "operationId": "ranking",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ranking found successfully"
                                        },
                                        "ranking": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "nickname": {
                                                        "type": "string"
                                                    },
                                                    "gameStats": {
                                                        "properties": {
                                                            "win_percentage": {
                                                                "type": "number"
                                                            },
                                                            "tie_percentage": {
                                                                "type": "number"
                                                            },
                                                            "lose_percentage": {
                                                                "type": "number"
                                                            },
                                                            "total_games": {
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No players found"
                                        },
                                        "ranking": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication",
                "name": "Bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}